Derivative Simplification and txt file addition
This commit contains modifications to the current differentiation code.

The code is a simiplification, as well as giving the user the ability to insert their equation in a single line, and write derivatives to whatever order they require.

A future update will include a substitution after the differentiation so that the user will be able to find the gradient at a specific points.